move scripts to hack and spellcheck
