+++ hack/check-fmt.sh
