+++ hack/test.sh
