+++ pkg/kritis/admission/admission.go
@@ -11,4 +11,4 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package admissionController
+package admission
